// src/api/generated/models/adminGetCatalogResponse.ts
/**
 * 由 Orval 根据 docs/openapi.yaml 自动生成；请勿手动修改。
 */
import type { CatalogDraft } from './catalogDraft';
import type { AdminGetCatalogResponseMessage } from './adminGetCatalogResponseMessage';

export interface AdminGetCatalogResponse {
  /**
   * 成功固定0
   * @minimum 0
   * @maximum 0
   */
  code: number;
  data: CatalogDraft;
  message: AdminGetCatalogResponseMessage;
}